Dados/pipelines/requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
dnspython==2.7.0
idna==3.10
numpy==2.3.0
matplotlib==3.9.0
seaborn==0.13.2
scikit-learn==1.5.0 
pandas==2.3.0
pymongo==4.13.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
six==1.17.0
tzdata==2025.2
urllib3==2.4.0
